Fix runtime import to determine new attribute types
