+++ yorm/bases/mappable.py
@@ -4,7 +4,6 @@ import abc
 import functools
 
 from .. import common
-from ..mapper import get_mapper
 
 
 log = common.logger(__name__)
@@ -22,7 +21,7 @@ def fetch_before(method):
     def wrapped(self, *args, **kwargs):
         """Decorated method."""
         if not _private_call(method, args):
-            mapper = get_mapper(self)
+            mapper = common.get_mapper(self)
             if mapper and mapper.modified:
                 log.debug("Fetching before call: %s", method.__name__)
                 mapper.fetch()
@@ -49,7 +48,7 @@ def store_after(method):
         result = method(self, *args, **kwargs)
 
         if not _private_call(method, args):
-            mapper = get_mapper(self)
+            mapper = common.get_mapper(self)
             if mapper and mapper.auto:
                 log.debug("Storing after call: %s", method.__name__)
                 mapper.store()
